feat: change Bulk for InteractionLagrangian
