+++ test/second_order.jl
@@ -1,0 +1,18 @@
+using KeldyshContraction, Test
+using KeldyshContraction: In, Out, Classical, Quantum, Plus, Minus
+using KeldyshContraction: Bulk, position
+
+@qfields ϕᶜ::Destroy(Classical) ϕᴾ::Destroy(Quantum)
+
+L_int =
+    im*(
+        0.5 * ϕᶜ' * ϕᴾ' * (ϕᶜ(Minus) * ϕᶜ(Minus) + ϕᴾ(Minus) * ϕᴾ(Minus)) -
+        0.5 * ϕᶜ(Plus) * ϕᴾ(Plus) * (ϕᶜ' * ϕᶜ' + ϕᴾ' * ϕᴾ') +
+        ϕᶜ' * ϕᴾ' * (ϕᶜ(Plus) * ϕᴾ(Plus) + ϕᶜ(Minus) * ϕᴾ(Minus))
+    )
+L = InteractionLagrangian(L_int)
+
+@testset "set_position" begin
+    @test position(L) == Bulk(0)
+    @test position(L(1)) == Bulk(1)
+end
